fixed #2487, #2495
